add Location#kilometres_to_dublin_office
This method calculates the distance between the lat and long on `self`
and the one from the Dublin office. We do this by using the
[first formula listed on
Wikipedia](https://en.wikipedia.org/wiki/Great-circle_distance#Formulae).

This PR also adds 2 tests for calculating the distance from London to
Dublin and Belfast to Dublin.